Only run test suite on Linux (skip Mac)
We seem to be having issues with memory limits on Mac when running this
test suite. So simply skip it for now. Running the test suite on Linux
is better than not running it at all. We can revisit testing on Mac once
we have a configuration of OpenBLAS that will work with the test suite.