On macOS skip lena image tests
As we were getting segfaults from OpenBLAS about memory again (this time
with the image tests), simply skip the tests that use the larger
`lena.png` image and stick to the tests using the smaller `boat.png`
test. Since these seem to pass locally with the `boat.png`, this seems
to confirm the theory that OpenBLAS is running into memory issues on
macOS and not something else.