Release 2.6
Bumps the version to 2.6. As there are some compatibility issues with
recent versions of NumPy, this tweaks some existing patches to work with
this newer release. Also some additional patches are needed for NumPy
compatibility to handle stricter type handling in NumPy. In particular
shapes that include non-integral values result in an exception. As the
Python 2 and 3 version of 2.6 come from different sources, we end up
with some duplication of patches to handle subtleties between the two
cases (e.g. different line numbers). Otherwise the patches are not too
drastic and some have already found their way upstream. So hopefully
these can be cutdown in future releases.