Drop BLAS symlink hack
This hack was added so that SPAMS would be able to pick up on OpenBLAS
being installed and link to it. However now that OpenBLAS has these
symlinks included in its package, it no longer makes sense to have this
hack in SPAMS. Further the fact that symlinks were being created on top
of existing files caused some build failures. So this provides only more
impetus to remove these hacks. Hence this commit does exactly that.