Use `defaults` compilers on macOS
As the `toolchain` doesn't quite fit our needs here (as we need `openmp`
on macOS), use the `defaults` compilers for macOS. This should provide
us `openmp` support on macOS and should help us start migrating this
package to the new compilers a bit earlier.

Should add we don't need to worry about C++ ABI breakage that occurs
when moving to the new compilers as this uses the C API of all
dependencies and the C++ code here is not exposed for linkage, but is
merely bound to the Python C API. For other parts of the conda-forge
stack, this is a concern.